Move BlockchainState to raiden_libs
